Expand drm definitions
